Fix sqrt_oram to allow for reshuffling.
In order for the sqrt_oram object to be updated (oprfs, indexes, stash, stash_count) the data needs to be modified in place, or it needs to be returned by the function. The latter seems to be converted by the SCALE-MAMBA combiler into a recursive call, which presents a problem since these can only have a certain depth.